[FLINK-8974] Run all-round DataSet job with failures in HA mode
